Chore: config로 빈 등록 주입
